refactor(tailwind docs): rename ClassNameReferences to UtilityClassReferences, add subfolders
