+++ docs/Systems/Tailwind/UtilityClassReferences/Backgrounds/BackgroundColor.stories.tsx
@@ -1,20 +1,21 @@
 import React from "react"
 import { Story } from "@storybook/react"
+import classnames from "classnames"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { CATEGORIES } from "../../../../storybook/constants"
-import { UtilityClassTemplate } from "../components/UtilityClassTemplate"
-import { flattenEntries } from "../helpers/flatten-entries"
-import { utilityDescription } from "../helpers/utilityDescription"
-import styles from "./styles.module.scss"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { flattenEntries } from "../../helpers/flatten-entries"
+import { utilityDescription } from "../../helpers/utilityDescription"
+import styles from "../styles.module.scss"
 
-const prefix = "border-"
+const prefix = "bg-"
 const classEntries = flattenEntries(
   prefix,
   kaizenTailwindTheme?.borderColor || {}
 )
 
 export default {
-  title: `${CATEGORIES.tailwind}/Classname References/Borders/Border Color`,
+  title: `${CATEGORIES.tailwind}/Utility Class References/Backgrounds/Background Color`,
   parameters: {
     docs: {
       description: {
@@ -28,12 +29,12 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
   isReversed,
 }) => (
   <UtilityClassTemplate
-    compiledCssPropertyName="border-color"
+    compiledCssPropertyName="background-color"
     classKeyValues={classEntries}
     renderExampleComponent={(cssProperty): React.ReactElement => (
       <div
-        style={{ borderColor: cssProperty }}
-        className={styles.boxWithBorder}
+        style={{ backgroundColor: cssProperty }}
+        className={classnames(styles.boxWithBorder, "rounded-default")}
       />
     )}
     isReversed={isReversed}
@@ -41,5 +42,5 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
 )
 
 export const StickerSheetDefault = StickerSheetTemplate.bind({})
-StickerSheetDefault.storyName = "Border Color"
+StickerSheetDefault.storyName = "Background Color"
 StickerSheetDefault.parameters = { chromatic: { disable: false } }
