+++ docs/Systems/Tailwind/UtilityClassReferences/Borders/BorderColor.stories.tsx
@@ -1,23 +1,20 @@
 import React from "react"
 import { Story } from "@storybook/react"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { CATEGORIES } from "../../../../storybook/constants"
-import { UtilityClassTemplate } from "../components/UtilityClassTemplate"
-import { utilityDescription } from "../helpers/utilityDescription"
-import styles from "./styles.module.scss"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { flattenEntries } from "../../helpers/flatten-entries"
+import { utilityDescription } from "../../helpers/utilityDescription"
+import styles from "../styles.module.scss"
 
 const prefix = "border-"
-const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
-  Object.entries(kaizenTailwindTheme?.borderWidth || []).map(
-    ([suffix, cssProperty]) => ({
-      utilityClassName: `${prefix}${suffix}`,
-      cssProperty,
-    })
-  )
+const classEntries = flattenEntries(
+  prefix,
+  kaizenTailwindTheme?.borderColor || {}
+)
 
 export default {
-  title: `${CATEGORIES.tailwind}/Classname References/Borders/Border Width`,
-  component: <div>Hello</div>,
+  title: `${CATEGORIES.tailwind}/Utility Class References/Borders/Border Color`,
   parameters: {
     docs: {
       description: {
@@ -31,11 +28,11 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
   isReversed,
 }) => (
   <UtilityClassTemplate
-    compiledCssPropertyName="border-width"
+    compiledCssPropertyName="border-color"
     classKeyValues={classEntries}
     renderExampleComponent={(cssProperty): React.ReactElement => (
       <div
-        style={{ borderWidth: cssProperty }}
+        style={{ borderColor: cssProperty }}
         className={styles.boxWithBorder}
       />
     )}
@@ -44,5 +41,5 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
 )
 
 export const StickerSheetDefault = StickerSheetTemplate.bind({})
-StickerSheetDefault.storyName = "Border Width"
+StickerSheetDefault.storyName = "Border Color"
 StickerSheetDefault.parameters = { chromatic: { disable: false } }
