+++ docs/Systems/Tailwind/UtilityClassReferences/Borders/BorderSpacing.stories.tsx
@@ -1,0 +1,54 @@
+import React from "react"
+import { Story } from "@storybook/react"
+import { kaizenTailwindTheme } from "@kaizen/tailwind"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { utilityDescription } from "../../helpers/utilityDescription"
+import styles from "../styles.module.scss"
+
+const prefix = "border-"
+const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
+  Object.entries(kaizenTailwindTheme?.spacing || []).map(
+    ([suffix, cssProperty]) => ({
+      utilityClassName: `${prefix}${suffix}`,
+      cssProperty,
+    })
+  )
+
+export default {
+  title: `${CATEGORIES.tailwind}/Utility Class References/Borders/Border Spacing`,
+  component: <div>Hello</div>,
+  parameters: {
+    docs: {
+      description: {
+        component: utilityDescription(prefix, classEntries[0].utilityClassName),
+      },
+    },
+  },
+}
+
+const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
+  isReversed,
+}) => (
+  <UtilityClassTemplate
+    compiledCssPropertyName="border-spacing"
+    classKeyValues={classEntries}
+    renderExampleComponent={(cssProperty): React.ReactElement => (
+      <table style={{ borderSpacing: cssProperty }}>
+        <tr>
+          <td className={styles.outlined}>Tutant</td>
+          <td className={styles.outlined}>Meenage</td>
+        </tr>
+        <tr>
+          <td className={styles.outlined}>Neetle</td>
+          <td className={styles.outlined}>Teetles</td>
+        </tr>
+      </table>
+    )}
+    isReversed={isReversed}
+  />
+)
+
+export const StickerSheetDefault = StickerSheetTemplate.bind({})
+StickerSheetDefault.storyName = "Border Spacing"
+StickerSheetDefault.parameters = { chromatic: { disable: false } }
