+++ docs/Systems/Tailwind/UtilityClassReferences/Effects/Shadow.stories.tsx
@@ -1,13 +1,14 @@
 import React from "react"
 import { Story } from "@storybook/react"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { CATEGORIES } from "../../../../storybook/constants"
-import { UtilityClassTemplate } from "../components/UtilityClassTemplate"
-import { utilityDescription } from "../helpers/utilityDescription"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { utilityDescription } from "../../helpers/utilityDescription"
+import styles from "../styles.module.scss"
 
-const prefix = "w-"
+const prefix = "shadow-"
 const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
-  Object.entries(kaizenTailwindTheme?.width || []).map(
+  Object.entries(kaizenTailwindTheme?.boxShadow || []).map(
     ([suffix, cssProperty]) => ({
       utilityClassName: `${prefix}${suffix}`,
       cssProperty,
@@ -15,7 +16,8 @@ const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
   )
 
 export default {
-  title: `${CATEGORIES.tailwind}/Classname References/Spacing/Width`,
+  title: `${CATEGORIES.tailwind}/Utility Class References/Effects/Box Shadow`,
+  component: <div>Hello</div>,
   parameters: {
     docs: {
       description: {
@@ -29,26 +31,15 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
   isReversed,
 }) => (
   <UtilityClassTemplate
-    compiledCssPropertyName="width"
+    compiledCssPropertyName="box-shadow"
     classKeyValues={classEntries}
     renderExampleComponent={(cssProperty): React.ReactElement => (
-      <div className="w-100 border-solid rounded-default h-100">
-        <div
-          className="h-100 flex items-center bg-blue-400"
-          style={{ width: cssProperty }}
-        >
-          {cssProperty.includes("content") ? (
-            <div className="font-family-paragraph bg-blue-100 p-4 border-dashed border-w-[1px]">
-              Inner content
-            </div>
-          ) : null}
-        </div>
-      </div>
+      <div style={{ boxShadow: cssProperty }} className={styles.box} />
     )}
     isReversed={isReversed}
   />
 )
 
 export const StickerSheetDefault = StickerSheetTemplate.bind({})
-StickerSheetDefault.storyName = "Width"
+StickerSheetDefault.storyName = "Box Shadow"
 StickerSheetDefault.parameters = { chromatic: { disable: false } }
