+++ docs/Systems/Tailwind/UtilityClassReferences/Modifiers/MediaQueries.stories.tsx
@@ -3,11 +3,11 @@ import { Story } from "@storybook/react"
 import { withDesign } from "storybook-addon-designs"
 import { Divider } from "@kaizen/draft-divider"
 import { Heading, Paragraph } from "@kaizen/typography"
-import { CATEGORIES } from "../../../../storybook/constants"
-import { figmaEmbed } from "../../../../storybook/helpers"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { figmaEmbed } from "../../../../../storybook/helpers"
 
 export default {
-  title: `${CATEGORIES.tailwind}/Classname References/Modifiers/Media Queries`,
+  title: `${CATEGORIES.tailwind}/Utility Class References/Modifiers/Media Queries`,
   parameters: {
     docs: {
       description: {
