+++ docs/Systems/Tailwind/UtilityClassReferences/Spacing/Height.stories.tsx
@@ -1,9 +1,9 @@
 import React from "react"
 import { Story } from "@storybook/react"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { CATEGORIES } from "../../../../storybook/constants"
-import { UtilityClassTemplate } from "../components/UtilityClassTemplate"
-import { utilityDescription } from "../helpers/utilityDescription"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { utilityDescription } from "../../helpers/utilityDescription"
 
 const prefix = "h-"
 const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
@@ -15,7 +15,7 @@ const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
   )
 
 export default {
-  title: `${CATEGORIES.tailwind}/Classname References/Spacing/Height`,
+  title: `${CATEGORIES.tailwind}/Utility Class References/Spacing/Height`,
   parameters: {
     docs: {
       description: {
