+++ docs/Systems/Tailwind/UtilityClassReferences/Typography/TextFontSize.stories.tsx
@@ -1,9 +1,9 @@
 import React from "react"
 import { Story } from "@storybook/react"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { CATEGORIES } from "../../../../storybook/constants"
-import { UtilityClassTemplate } from "../components/UtilityClassTemplate"
-import { utilityDescription } from "../helpers/utilityDescription"
+import { CATEGORIES } from "../../../../../storybook/constants"
+import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { utilityDescription } from "../../helpers/utilityDescription"
 
 const prefix = "text-"
 const classEntries = Object.entries(kaizenTailwindTheme?.fontSize || []).map(
@@ -14,7 +14,7 @@ const classEntries = Object.entries(kaizenTailwindTheme?.fontSize || []).map(
 )
 
 export default {
-  title: `${CATEGORIES.tailwind}/Classname References/Typography/Font Size`,
+  title: `${CATEGORIES.tailwind}/Utility Class References/Typography/Font Size`,
   parameters: {
     docs: {
       description: {
